PDF object numbers need not be int64_t, int is sufficient.
This is true because they are now limited below PDF_MAX_OBJECT_NUMBER.

+++ include/mupdf/pdf/xref.h
@@ -58,7 +58,7 @@ struct pdf_xref_subsec_s
 {
 	pdf_xref_subsec *next;
 	int len;
-	int64_t start;
+	int start;
 	pdf_xref_entry *table;
 };
